requirements.txt
ultralytics
opencv-python
pyyaml
matplotlib
numpy
fpdf2
tqdm
pandas
